+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.32.1](https://github.com/jalbarrang/torena-sim/compare/v0.32.0...v0.32.1) (2026-07-18)
 
+### Features
+
+* **carat:** correct event, login, Training Pass, daily-pack, and League of Heroes income; show all-in monthly income; add reference spreadsheet differential coverage
+
 ## [0.32.0](https://github.com/jalbarrang/torena-sim/compare/v0.31.3...v0.32.0) (2026-07-18)
 
 ## [0.31.3](https://github.com/jalbarrang/torena-sim/compare/v0.31.2...v0.31.3) (2026-07-17)